added check for pip

+++ setup.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-curl https://bootstrap.pypa.io/get-pip.py > get-pip.py
-sudo python get-pip.py
+if [[ -z `type -p pip` ]]; then
+    curl https://bootstrap.pypa.io/get-pip.py > get-pip.py
+    sudo python get-pip.py
+fi
 
 pip install virtualenv
 virtualenv -p python3.6 venv
@@ -10,4 +12,7 @@ source venv/bin/activate
 pip install -r requirements.txt
 
 deactivate 
-rm get-pip.py
+
+if [[ -a get-pip.py ]]; then
+    rm get-pip.py
+fi
